# public/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/consultants/{id}/availabilities": {
      "get": {
        "operationId": "ConsultantsController_findAllAvailabilities",
        "summary": "Get all availabilities for a consultant",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Consultant ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the consultant availabilities",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Availability"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Consultant not found"
          }
        },
        "tags": [
          "consultants"
        ]
      }
    },
    "/consultants/availabilities/{id}": {
      "get": {
        "operationId": "ConsultantsController_findAvailabilityById",
        "summary": "Get an availability by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Availability ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the availability",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Availability"
                }
              }
            }
          },
          "404": {
            "description": "Availability not found"
          }
        },
        "tags": [
          "consultants"
        ]
      },
      "patch": {
        "operationId": "ConsultantsController_updateAvailability",
        "summary": "Update an availability",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Availability ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAvailabilityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Availability updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Availability"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Availability not found"
          }
        },
        "tags": [
          "consultants"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "ConsultantsController_removeAvailability",
        "summary": "Delete an availability",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Availability ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Availability deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Availability not found"
          }
        },
        "tags": [
          "consultants"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/consultants/availabilities": {
      "post": {
        "operationId": "ConsultantsController_createAvailability",
        "summary": "Create a new availability",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAvailabilityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Availability created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Availability"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "409": {
            "description": "Time slot overlaps with existing availability"
          }
        },
        "tags": [
          "consultants"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/consultants/{id}/available-slots": {
      "get": {
        "operationId": "ConsultantsController_getAvailableTimeSlots",
        "summary": "Get available time slots for a consultant on a specific date",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Consultant ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": true,
            "in": "query",
            "description": "Date in YYYY-MM-DD format",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the available time slots",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Availability"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Consultant not found"
          }
        },
        "tags": [
          "consultants"
        ]
      }
    },
    "/consultants": {
      "get": {
        "operationId": "ConsultantsController_findAll",
        "summary": "Get all consultants with optional filtering",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "specialties",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "isVerified",
            "required": false,
            "in": "query",
            "schema": {
              "default": true,
              "type": "boolean"
            }
          },
          {
            "required": false,
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "required": false,
            "name": "specialties",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "required": false,
            "name": "isVerified",
            "in": "query",
            "schema": {
              "default": true,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a list of consultants",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Consultant"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "consultants"
        ]
      },
      "post": {
        "operationId": "ConsultantsController_create",
        "summary": "Create a new consultant profile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConsultantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Consultant created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Consultant"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "409": {
            "description": "User is already a consultant"
          }
        },
        "tags": [
          "consultants"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/consultants/{id}": {
      "get": {
        "operationId": "ConsultantsController_findOne",
        "summary": "Get a consultant by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Consultant ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the consultant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Consultant"
                }
              }
            }
          },
          "404": {
            "description": "Consultant not found"
          }
        },
        "tags": [
          "consultants"
        ]
      },
      "patch": {
        "operationId": "ConsultantsController_update",
        "summary": "Update a consultant profile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Consultant ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateConsultantDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Consultant updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Consultant"
                }
              }
            }
          },
          "404": {
            "description": "Consultant not found"
          }
        },
        "tags": [
          "consultants"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "ConsultantsController_remove",
        "summary": "Delete a consultant",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Consultant ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Consultant deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Consultant not found"
          }
        },
        "tags": [
          "consultants"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/consultants/{id}/verify": {
      "patch": {
        "operationId": "ConsultantsController_verify",
        "summary": "Verify a consultant",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Consultant ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "isVerified": {
                    "type": "boolean",
                    "description": "Verification status",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Consultant verification updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Consultant"
                }
              }
            }
          },
          "404": {
            "description": "Consultant not found"
          }
        },
        "tags": [
          "consultants"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
    "/sessions/consultant/{consultantId}": {
      "get": {
        "operationId": "SessionsController_findByConsultant",
        "summary": "Get sessions by consultant ID",
        "parameters": [
          {
            "name": "consultantId",
            "required": true,
            "in": "path",
            "description": "Consultant ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns sessions for the specified consultant",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Session"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Consultant not found"
          }
        },
        "tags": [
          "sessions"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
    "/reviews/consultant/{consultantId}": {
      "get": {
        "operationId": "ReviewsController_findByConsultant",
        "summary": "Get reviews by consultant ID",
        "parameters": [
          {
            "name": "consultantId",
            "required": true,
            "in": "path",
            "description": "Consultant ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns reviews for the specified consultant",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Review"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Consultant not found"
          }
        },
        "tags": [
          "reviews"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
    "/admin/consultants/{id}/verify": {
      "patch": {
        "operationId": "AdminController_verifyConsultant",
        "summary": "Verify or reject a consultant",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Consultant ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "isVerified"
                ],
                "properties": {
                  "isVerified": {
                    "type": "boolean",
                    "description": "Verification status",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Consultant verification status updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Consultant"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Requires admin role"
          },
          "404": {
            "description": "Consultant not found"
          }
        },
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "JWT-auth": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Family Counseling API",
    "description": "\n      # Family Counseling API Documentation\n      \n      This API provides the backend services for the Family Counseling Platform.\n\n      ## Authorization\n      Most endpoints require JWT authentication. After logging in, include the JWT token in the Authorization header:\n      ```\n      Authorization: Bearer YOUR_JWT_TOKEN\n      ```\n      \n      ## Role-based access\n      Different endpoints have different role requirements. The API supports these roles:\n      - admin: Full access to all endpoints\n      - consultant: Access to consultant-specific features\n      - client: Access to client-specific features\n    ",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "ApiResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Status of the operation",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "Response message",
            "example": "Operation successful"
          },
          "data": {
            "type": "object",
            "description": "Response data",
            "example": {}
          },
          "error": {
            "type": "string",
            "description": "Error description if any",
            "example": null
          }
        },
        "required": [
          "success",
          "message",
          "data",
          "error"
        ]
      },
      "Availability": {
        "type": "object",
        "properties": {
          "startTime": {
            "format": "date-time",
            "type": "string",
            "description": "Start time of availability"
          },
          "endTime": {
            "format": "date-time",
            "type": "string",
            "description": "End time of availability"
          },
          "isAvailable": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "startTime",
          "endTime",
          "isAvailable"
        ]
      },
      "CreateAvailabilityDto": {
        "type": "object",
        "properties": {
          "consultantId": {
            "type": "string"
          },
          "startTime": {
            "type": "string",
            "description": "Start time of availability"
          },
          "endTime": {
            "type": "string",
            "description": "End time of availability"
          },
          "isAvailable": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "consultantId",
          "startTime",
          "endTime",
          "isAvailable"
        ]
      },
      "UpdateAvailabilityDto": {
        "type": "object",
        "properties": {
          "startTime": {
            "type": "string"
          },
          "endTime": {
            "type": "string"
          },
          "isAvailable": {
            "type": "boolean"
          }
        }
      },
      "Consultant": {
        "type": "object",
        "properties": {
          "specialties": {
            "description": "Consultant specialties",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "bio": {
            "type": "string",
            "description": "Consultant biography"
          },
          "education": {
            "type": "string",
            "description": "Education and qualifications"
          },
          "rating": {
            "type": "number",
            "default": 0
          },
          "reviewCount": {
            "type": "number",
            "default": 0
          },
          "isVerified": {
            "type": "boolean",
            "default": false
          },
          "consultantLicense": {
            "type": "string"
          }
        },
        "required": [
          "specialties",
          "bio",
          "education",
          "rating",
          "reviewCount",
          "isVerified"
        ]
      },
      "CreateConsultantDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "specialties": {
            "description": "List of specialties",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "bio": {
            "type": "string",
            "description": "Consultant biography"
          },
          "education": {
            "type": "string",
            "description": "Education and qualifications"
          },
          "consultantLicense": {
            "type": "string"
          }
        },
        "required": [
          "userId",
          "specialties",
          "bio",
          "education"
        ]
      },
      "UpdateConsultantDto": {
        "type": "object",
        "properties": {
          "specialties": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "bio": {
            "type": "string"
          },
          "education": {
            "type": "string"
          },
          "consultantLicense": {
            "type": "string"
          }
        }
      },
      "Client": {
        "type": "object",
        "properties": {}
      },
...
      "SessionStatus": {
        "type": "string",
        "description": "Current status of the session",
        "default": "pending",
        "enum": [
          "pending",
          "confirmed",
          "completed",
          "cancelled"
        ]
      },
      "Session": {
        "type": "object",
        "properties": {
          "consultant": {
            "description": "The consultant providing the session",
            "allOf": [
              {
                "$ref": "#/components/schemas/Consultant"
              }
            ]
          },
          "client": {
            "description": "The client receiving the session",
            "allOf": [
              {
                "$ref": "#/components/schemas/Client"
              }
            ]
          },
          "date": {
            "format": "date-time",
            "type": "string",
            "description": "Session date and time",
            "example": "2025-04-15T14:00:00Z"
          },
          "status": {
            "example": "confirmed",
            "$ref": "#/components/schemas/SessionStatus"
          },
          "notes": {
            "type": "string",
            "description": "Session notes or summary",
            "example": "Client had concerns about parenting strategies. Discussed positive reinforcement techniques."
          },
          "messengerId": {
            "type": "string",
            "description": "External messenger identifier (e.g., Telegram chat ID)",
            "example": "12345678"
          },
          "messengerType": {
            "type": "string",
            "description": "Type of messenger used for communication",
            "example": "telegram",
            "enum": [
              "telegram",
              "whatsapp"
            ]
          }
        },
        "required": [
          "consultant",
          "client",
          "date",
          "status"
        ]
      },
...
      "Review": {
        "type": "object",
        "properties": {
          "rating": {
            "type": "number",
            "description": "Rating from 1 to 5",
            "minimum": 1,
            "maximum": 5
          },
          "comment": {
            "type": "string",
            "description": "Review comment"
          }
        },
        "required": [
          "rating",
          "comment"
        ]
      },
...
    }
  }
}